refactor!: amend TLE and Satellite classes in satelles module
refactor!: amend TLE and Satellite classes in satelles module
